Removed dependency on monsters.txt and static data
