+++ bin/dev-server.js
@@ -56,7 +56,6 @@ async function doIt() {
 
   // dump monsters.txt
   var monstersDB = new PouchDB('http://localhost:6984/monsters');
-  var monstersFullFatDB = new PouchDB('http://localhost:6984/monsters-fullfat');
   var descriptionsDB = new PouchDB('http://localhost:6984/descriptions');
   var evolutionsDB = new PouchDB('http://localhost:6984/evolutions');
   var typesDB = new PouchDB('http://localhost:6984/types');
@@ -70,7 +69,6 @@ async function doIt() {
     evolutionsDB.load(await fs.readFileAsync('src/assets/evolutions.txt', 'utf-8')),
     typesDB.load(await fs.readFileAsync('src/assets/types.txt', 'utf-8')),
     movesDB.load(await fs.readFileAsync('src/assets/moves.txt', 'utf-8')),
-    monstersFullFatDB.load(await fs.readFileAsync('src/assets/monsters.txt', 'utf-8')),
     monsterMovesDB.load(await fs.readFileAsync('src/assets/monster-moves.txt', 'utf-8')),
     monstersSupplementalDB.load(await fs.readFileAsync('src/assets/monsters-supplemental.txt', 'utf-8'))
   ];
